pyproject.toml
[build-system]
requires = ["uv_build >= 0.9.21, <0.10.0"]
build-backend = "uv_build"

[project]
name = "gtasks-cli"
version = "0.1.0"
description = "Google Tasks Command Line Interface"
readme = "README.md"
requires-python = ">= 3.14"
license = { text = "MIT" }
authors = [
    { name = "Nikita Filippov" },
]
dependencies = [
    "google-auth>=2.41",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.187",
    "bidict>=0.23.1",
    "pytest-mock>=3.15.1",
    "google-api-python-client-stubs>=1.31.0",
    "ty>=0.0.9",
]

[project.scripts]
gtasks = "gtasks.cli:main"

# uv looks in src/gtasks_cli/__init__.py by default
[tool.uv.build-backend]
module-root = ""
module-name = "gtasks"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
